bugfix
Bernula Szonja
